Update login procedure to support multiple methods
Now, you can sign in either with password or email. Also, there
are nice pages for everything that occurs, like "email sent" or
"account inactive".

Also, a signin_mode has been added to the User model to decide how
that users wants to sign in.Email sign-in is yet to beset up, though.